api/requirements.txt
channels[daphne]==4.2.0
django==5.1.6
django-phonenumber-field[phonenumberslite]==8.0.0
djangorestframework==3.15.2
pillow==11.1.0